Better status logging and other minor enhancements

+++ setup.sh
@@ -1,44 +1,69 @@
 #!/bin/bash
 
-# install homebrew
+echo "\n\n****installing homebrew****"
 /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 
+echo "\n\n****installing python****"
 brew install python
 
-# install oh-my-zsh
+echo "\n\n****installing oh-my-zsh****"
 sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 
-# copy zshrc
+echo "\n\n****copying zshrc****"
 curl -fsSL -o ~/.zshrc https://raw.githubusercontent.com/invacuo/dev-setup/master/.zshrc
 
-# install powerline fonts
-git clone https://github.com/powerline/fonts.git
-. ~/fonts/install.sh
-rm -rf ~/fonts.sh
+echo "\n\n****installing powerline fonts****"
+if [ -d ~/fonts ]; then
+  while true; do
+    echo "Fonts directory already exists. Do you want to delete it $1? (y/n): "
+    read yn
+    case $yn in
+      [Yy]* )
+        echo "Deleting ~/fonts"
+        rm -rf ~/fonts
+        git clone https://github.com/powerline/fonts.git ~/fonts
+        . ~/fonts/install.sh
+        rm -rf ~/fonts.sh
+        break
+        ;;
+      [Nn]* )
+        echo "Skipping fonts installation. Powerline status may not work correctly."
+        break
+        ;;
+      * )
+        echo "Please answer yes or no."
+        ;;
+    esac
+  done
+fi
 
-#install powerline status
+echo "\n\n****installing powerline status****"
 pip3 install --user powerline-status
 
-#install hub for doing github stuff without going to github
+echo "\n\n****installing hub for doing github stuff without going to github****"
 brew install hub
+
+echo "\n\n****installing tmux****"
 brew install tmux
 
-#install firacode
-brew tap caskroom/fonts
+echo "\n\n****installing firacode****"
+brew tap homebrew/cask-fonts
 brew cask install font-fira-code
 
-# copy personal shortcuts and add it to .zshrc
+echo "\n\n****copying personal shortcuts and add it to .zshrc****"
 curl -fsSL -o ~/shortcuts.sh https://raw.githubusercontent.com/invacuo/git-shortcuts/master/shortcuts.sh
 
-# install atom-packages
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/invacuo/dev-setup/master/atom-package-installer.sh)"
-
+echo "\n\n****installing rbenv****"
+brew install rbenv
 echo 'export PATH="$HOME/.rbenv/bin:$PATH"' >> ~/.zshrc
+echo 'eval "$(rbenv init -)"' >> ~/.zshrc
 
 #git config
 git config --global push.default current
 
-# vim stuff
+echo "\n\n****doing vim stuff like setting colors and fonts and copying vimrc****"
 mkdir ~/.vim ~/.vim/colors
 curl -o ~/.vim/colors/monokai.vim https://raw.githubusercontent.com/sickill/vim-monokai/master/colors/monokai.vim
 curl https://raw.githubusercontent.com/invacuo/dev-setup/master/.vimrc >> ~/.vimrc
+
+echo "\n\n****DONE****"
